refactor(metadata): move display name and overlay functions to metadata module
